app/front/requirements.txt
streamlit==1.29.0
requests==2.28.2
python-dotenv==1.0.0
google-generativeai==0.3.2
numpy==1.24.4
pandas==1.5.3
requests==2.28.2
python-dotenv==1.0.0
google-generativeai==0.3.2
loguru==0.6.0